refactor: isolate procfs file server adapter
